maintain the mesos master url in bootstrap.py
Prior to this commit, the Mesos master URL in Jenkins' config.xml would
be set at package install time and then never touched again. If we need
to change this across versions of the package, this would require manual
intervention by our users. That isn't a great experience.

This commit modifies the value of 'master' each time bootstrap.py runs,
which is whenever the container is launched. It will set this value to
the one provided in the script, so that it can be realistically changed
between releases (although we don't anticipate needing to do this
often). This value will then be provided via the package's config.json
file in the Mesosphere Universe.


Former-commit-id: c9cec4dcc706e52dd19ffdd7683b8f1f1630a2be