populate ssh known_hosts file on the master
This commit introduces a function that runs `ssh-keyscan` on a string of
hostnames and/or IP addresses and retrieves the SSH public keys for
them. It then writes this output into /root/.ssh/known_hosts in the
container that runs the Jenkins master.

This is part of bootstrap.py so that users can specify a list of hosts,
internal and/or external, when they install the Jenkins package for
DCOS.


Former-commit-id: b0ead4cad19eb2cbff89eb9f2b2fd0edd14db0a4