Rename Local Variable and Set Default
Change the local variable from jenkins_instance_url to jenkins_root_url
to avoid confusion and be consistent with the environment variable.

Leverage the "default" parameter of os.getenv to set the root URL to
the Marathon host and Nginx port if the environment variable is not
set. This avoids checking for None later.