SSL Vars Should Be Absolute Paths
The passed in SSL paths are relative to the MESOS_SANDBOX for DC/OS 1.8
and below (fix intended for 1.9). This is not the directory where Jenkins
is executed from. This commit adds a script that transforms the relative
to absolute paths and allows Jenkins to start on a strict cluster.