Create Dockerfile From Official Jenkins Docker Image
The Dockerfile is now based off of the official Jenkins docker
image and handles plugin installation and upgrading in a different
manner.

The bootstrap script had to be altered as the first run criteria
has changed from the folder itself missing to the config.xml not
yet being written to disk.