Thread Improvements (#235)
* Add thread timeout

* Add documentation and break scale creation into more methods

Add documentation throughout the scale tool, both at the function
and line level.

The initial function to create scale workloads is too big and
this begins to break some of the functionality out into smaller
functions.

* Capture result from thread runs

Add custom thread type that stores result of run() so that
failures can be detected and reported. This is used to
determine what Jenkins instances had issues and can be skipped
in future steps.

Move timeout and error detection into its own function so both
setup and teardown can use it. Any unsuccessful scenarios will be
reported through the logger at the warning level.

* Creating jobs is now threaded